Fixing imports, moving helpers around
